Validate job class before delayed enqueue.
Resque does check of the job class before enqueue it.
Resque-scheduler doesn't do that before put them into the schedule.
That is why we see error no on schedule but at execution.
This is not correct. We should do same checks as redis do:
https://github.com/defunkt/resque/blob/master/lib/resque/job.rb#L39

Unfortunatelly we can not reuse code.