Support enqueued jobs that hate Resque::Job.create
Queuing Resque::JobWithStatus jobs fail, as they fall back to the normal Resque::Job.create behavior when they are scheduled.  This is because the job_hash doesn't include the :custom_job_class parameter that the config file has support for.

This change allows the jobs to set an instance variable (which isn't the best way but works) that indicates they require the custom_job_class parameter.