Use job name as implicit class name in the schedule
When configuring the schedule, the only required argument is 'class'.
On the other hand, the job name (hash key) is not too useful and usually
just duplicates the resque worker class name (underscored).

So instead of having to do:

    some_job:
      class: SomeJob
      cron: '* * * * *'
    another_job:
      class: AnotherJob
      every: 1m
    job_name_different_than_class:
      class: WeirdlyNamedClass

You can now do:

    SomeJob:
      cron: '* * * * *'
    AnotherJob:
      every: 1m
    job_name_different_than_class:
      class: WeirdlyNamedClass

By default, the job name will be treated as class name,
but you can still provide a different job name and class name.