Change recommended way of loading schedule
When schedule is loaded in resque:setup task, all resque process irrespective of if they are the scheduler or not will try to reload the schedule. 

Not only is this an unnecessary overhead, but it also causes a race condition in setting the schedule which was fixed recently in #439. Since this fix is not yet available on rubygems, it best to tell user to only run one instance of the scheduler and do it like this.