Use `Resque.enqueue_to` to retain enqueue behaviour
If we create the resque job directly, we skip the `before_enqueue` and `after_enqueue` callbacks. This causes surprisingly inconsistent behaviour.

This change mirrors `lib/resque/scheduler.rb:272`.

This ensures consistent behaviour for the follow scenarios: (1) when scheduling something for the future, (2) scheduling something for the future and `Resque.inline == true` (i.e., run it immediately), (3) scheduleing something for the past (i.e., run it immediately).