Ensure that jobs are directly created for jobs enqueued at Time.now
If less than a second passes between calling Resque.enqueue_in(0, ...),
which is not all that unlikely, a job may be delayed when we don't
expect the job to be delayed.  In high traffic applications where the
resque scheduler process is very busy, this unexpected delay can grow as
other jobs are being enqueued after their own delays.  Since
resque-scheduler only supports multiple instances for redundancy, there
is no way to scale resque-scheduler horizontally.  Thus my goal is to ensure
that as many jobs as possible are created directly and not unexpectedly
delayed.