Refactor remove_delayed_selection to allow massive jobs cleanup
The need for this change arised when we had incident that caused scheduling of 700,000 invalid jobs that had to be cleaned.
This method deletes the jobs as it finds them, so if the process is interrupted it can be resumed.

Here's why I didn't use the existing methods:
`find_delayed_selection` - would load all the matching jobs in memory, high memory usage and impossible to track progress of the scanning
`remove_delayed_job` - is inconvenient as it tries to delete the job from all timestamps, which is not needed here, because it goes over all timestamps anyway
`remove_delayed_job_from_timestamp` seemed to do what's needed but it expects decoded arguments and reencodes them, which is unnecessary computation and it calls `clean_up_timestamp`, which previously wasn't done as part of this method.